+++ src/main/java/bob/AddCommand.java
@@ -1,4 +1,4 @@
-package Bob;
+package bob;
 
 import java.io.IOException;
