+++ src/main/java/bob/BobException.java
@@ -1,4 +1,4 @@
-package Bob;
+package bob;
 
 public class BobException extends Exception {
 }
