+++ src/main/java/bob/Command.java
@@ -1,4 +1,4 @@
-package Bob;
+package bob;
 
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage);
