+++ src/main/java/bob/Deadline.java
@@ -1,4 +1,4 @@
-package Bob;
+package bob;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
