+++ src/main/java/bob/DeleteCommand.java
@@ -1,4 +1,4 @@
-package Bob;
+package bob;
 
 import java.io.IOException;
