+++ src/main/java/bob/EmptyFieldException.java
@@ -1,4 +1,4 @@
-package Bob;
+package bob;
 
 public class EmptyFieldException extends BobException {
 }
