+++ src/main/java/bob/HelpCommand.java
@@ -1,4 +1,4 @@
-package Bob;
+package bob;
 
 public class HelpCommand extends Command {
     @Override
