+++ src/main/java/bob/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package Bob;
+package bob;
 
 public class InvalidCommandException extends BobException {
