+++ src/main/java/bob/Storage.java
@@ -1,4 +1,4 @@
-package Bob;
+package bob;
 
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
