+++ src/main/java/bob/ToDo.java
@@ -1,4 +1,4 @@
-package Bob;
+package bob;
 
 public class ToDo extends Task {
     public ToDo(String description) {
