+++ src/main/java/bob/Ui.java
@@ -1,4 +1,4 @@
-package Bob;
+package bob;
 
 import java.util.Scanner;
