+++ text-ui-test/runtest.bat
@@ -7,7 +7,7 @@ REM delete output from previous run
 if exist ACTUAL.TXT del ACTUAL.TXT
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java\Bob -Xlint:none -d ..\bin ..\src\main\java\Bob\*.java
+javac  -cp ..\src\main\java\bob -Xlint:none -d ..\bin ..\src\main\java\bob\*.java
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
     exit /b 1
@@ -15,7 +15,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin Bob < input.txt > ACTUAL.TXT
+java -classpath ..\bin bob < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
